- Aider: corrigé des problèmes subtils avec la fusion des personnes.
git-svn-id: https://svn.opac.ch/svn/cr/branches/vs2010.1@22475 787ff09c-d2b5-7842-b444-8dd0a418631d

+++ Epsitec.Cresus/Aider/Entities/AiderContactEntity.cs
@@ -315,7 +315,7 @@ namespace Epsitec.Aider.Entities
 			}
 		}
 
-		public static void Delete(BusinessContext businessContext, AiderContactEntity contact)
+		public static void Delete(BusinessContext businessContext, AiderContactEntity contact, bool deleteParticipations = true)
 		{
 			if (contact.ContactType == ContactType.PersonAddress)
 			{
@@ -340,7 +340,10 @@ namespace Epsitec.Aider.Entities
 				legalPerson.RemoveContactInternal (contact);
 			}
 
-			AiderContactEntity.DeleteParticipations (businessContext, contact);
+			if (deleteParticipations)
+			{
+				AiderContactEntity.DeleteParticipations (businessContext, contact);
+			}
 			
 			businessContext.DeleteEntity (contact);
 		}
